Use the `-` flag for checking STDIN
PHPCS offers the `-` flag on the end of the command and params to make it explicit, that we are going to read from STDIN instead of local files and directories. While omitting the `-` often works, it does not seem to work with empty files as our `cat empty.file | phpcs --report=json` seems to be hang in an infinite loop wainting for input data.

Adding `-` flag fixes the issue.