Force quiet mode (#8)
While the quiet mode in PHPCS is the default setting, it can be changed via `show_progress` config. Turning the `show_progress` on / disabling the quiet mode, makes the PHPCS to not only output the JSON contianing warnings and errors, but also a progress on individual files (eg.: `E 1 / 1 (100%)`), which then breaks the phpcs-changed's JSON parsing.

By using the `-q` flag in the `phpcs --format=json` commands, we would force the quiet mode for phpcs-changed run commands only, not affecting local settings, and making sure the phpcs-changed works no matter what the local setting is.