Add support for remote branch diff in git workflow
This feature is meant for running the PHPCS in between two git branches, the current one and a specified one (using the new --git-branch param), taking advantage of the `git diff branch...` command, where `branch` is a remote branch (eg.: `master`).

In order to accommodate such diff in git workflow, the UnixShell had to be updated the way it uses `exec` instead of `shell_exec`, as checking an existence of a file in remote branch can't be done via `git status`, but rather via `git cat-file -e`, where the return status is what tells us whether the file/object exists or not.