Add arc-lint option and improve some git performance (#49)
* Check on whether the command is run from git root

Instead of bluntly using the `git ls-files --full-name` every  time `git
show` is called, we can check on whether the PhpcsChanges is run from
git's root and if so, we can use the passed in file name w/o additional
check which should save some precious time and speed up the command.

* Only calculate the old and new file git object hash when cache is used

Currently both, the old and new git file object hash is being calculated
even in case the cache is not enabled, which introduces some git
commands overhead for computing something which is not being used at
all.

* Add arc-lint flag to the PhpcsChanged command

The flag allow the PhpcsChanged command to be aware of being run via the `arc lint` command and perform some performance improvements, eg.: skipping the `validateGitFileExists` check, which is being performed by arcanist anyway.

The check is slow at times and we should see some performance improvement on WPCOM.

* Use `--porcelain` flag instead of `--short` for `git status` used in validateGitFileExists

The output of `--porcelain` is similar/same to the one of `--short`, just is intended for usage in scripts and is guaranteed to not change in future versions of git. It should make the script more robust.

* Silence the fatal error of `git cat-file -e` command

In case the file does not exist, a fatal error is being displayed, which does not feel appropriate and may confuse users. Nothing actually happens, just the file does not exists in original branch and is being linted in a standard way.

* Fix tests after changing the `git status --short` to `git status --porcelain`

* Replace one more instance of `git status --short` with `--porcelain`

* Check on whether the old and new file hash was calcualted before attempting to use it in debug message.

Addresses PHP notices for undefined variable.

* Register the newly added `git rev-parse --show-toplevel` command in tests.

* Fix a newly introduced typo s/porelain/porcelain/

* Address phpstan issues around possibly undefined newFileHash and oldFileHash variables.

Always define those, yet as empty strings + fix relace the `isset()` check for `!empty()` one.

* Remove extra single quote.

Co-authored-by: Payton Swick <payton@foolord.com>

* Remove extra single quote.

Co-authored-by: Payton Swick <payton@foolord.com>

* Remove unintentionally introduced newline.

* Add the `--arc-lint` option to README.

Co-authored-by: Payton Swick <payton@foolord.com>