Optimize variables defined in the catch statement.
Do not set the `$newFilePhpcsOutput` variable in catch statement as it's not actually being used later in the code.

In this commit, the value of `$newFilePhpcsMessages` variable is being calculated from `PhpcsMessages::fromPhpcsJson` method with an empty string passed in as param, since it's how it's calculated for the `$oldFilePhpcsOutput`.